fix: refactoring, adding translations
